tools/python/requirements.txt
PyYAML
semver
birdparse
birdyaml
birdjson
birdpy
zeroconf
pyserial
websocket-client
esptool
plac
